+++ examples/async/index.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import App from './containers/App';
+import Root from './containers/Root';
+import BrowserHistory from 'react-router/lib/BrowserHistory';
 
 React.render(
-  <App />,
+  <Root history={new BrowserHistory()} />,
   document.getElementById('root')
 );
