+++ examples/async/package.json
@@ -29,6 +29,7 @@
   "dependencies": {
     "react": "^0.13.3",
     "react-redux": "^0.8.0",
+    "react-router": "^1.0.0-beta3",
     "redux": "^1.0.0-rc",
     "redux-thunk": "^0.1.0"
   },
